<refactoring> command ssd 파일 분리
